Fix compiler warning (refs #7088)
git-svn-id: https://svn.osgeo.org/gdal/trunk@40441 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_vsil_curl.cpp
@@ -1149,7 +1149,7 @@ retry:
                 eExists = EXIST_YES;
                 fileSize = CPLScanUIntBig(
                     pszContentLength,
-                    strlen(pszContentLength));
+                    static_cast<int>(strlen(pszContentLength)));
                 if( ENABLE_DEBUG )
                     CPLDebug("VSICURL", "GetFileSize(%s)=" CPL_FRMT_GUIB,
                             osURL.c_str(), fileSize);
